Move persona hooks to catalog personas
