+++ docs/developer/modules.html
@@ -99,10 +99,10 @@
             </tr>
             <tr>
               <td><code>features/catalog/personas</code></td>
-              <td>Persona panel/editor/modals, imports, and query keys.</td>
+              <td>Persona panel/editor/modals, imports, data hooks, and query keys.</td>
               <td>
-                Storage API, avatar API, persona contracts. Current shared persona data hooks are exported from
-                <code>features/catalog/characters/index.ts</code>.
+                Storage API, avatar API, persona contracts. Shared persona data hooks are exported from
+                <code>features/catalog/personas/index.ts</code>.
               </td>
             </tr>
             <tr>
