+++ src/features/catalog/characters/hooks/use-characters.ts
@@ -1,5 +1,5 @@
 // ──────────────────────────────────────────────
-// React Query: Character, Group & Persona hooks
+// React Query: Character & Group hooks
 // ──────────────────────────────────────────────
 import { useMemo } from "react";
 import { useQuery, useQueries, useMutation, useQueryClient, type QueryClient } from "@tanstack/react-query";
@@ -15,7 +15,6 @@ import { ApiError } from "../../../../shared/api/api-errors";
 import { storageApi } from "../../../../shared/api/storage-api";
 import { storageCommandsApi } from "../../../../shared/api/storage-commands-api";
 import { galleryApi, spriteApi } from "../../../../shared/api/image-generation-api";
-import { personaApi } from "../../../../shared/api/persona-api";
 import type { CharacterCardVersion } from "../../../../engine/contracts/types/character";
 import type { SpriteCapabilities, SpriteCleanupEngine } from "../../../../shared/types/sprite-capabilities";
 
@@ -40,21 +39,6 @@ export type CharacterSummary = {
   updatedAt?: string;
 };
 
-export type PersonaSummary = {
-  id: string;
-  name?: string;
-  comment?: string | null;
-  description?: string;
-  tags?: string[];
-  avatarPath?: string | null;
-  avatarCrop?: unknown;
-  isActive?: string | boolean;
-  createdAt?: string;
-  nameColor?: string;
-  dialogueColor?: string;
-  boxColor?: string;
-};
-
 const CHARACTER_LIST_FIELDS = ["id", "data", "comment", "avatarFilePath", "avatarFilename", "createdAt", "updatedAt"];
 
 const CHARACTER_SUMMARY_OPTIONS = {
@@ -66,24 +50,6 @@ const CHARACTER_LIST_OPTIONS = {
 };
 const CHARACTER_SUMMARY_BY_ID_CONCURRENCY = 8;
 const EMPTY_CHARACTER_SUMMARIES: CharacterSummary[] = [];
-
-const PERSONA_SUMMARY_OPTIONS = {
-  fields: [
-    "id",
-    "name",
-    "comment",
-    "description",
-    "tags",
-    "avatarPath",
-    "avatarCrop",
-    "isActive",
-    "active",
-    "createdAt",
-    "nameColor",
-    "dialogueColor",
-    "boxColor",
-  ],
-};
 
 function isCharacterListRecord(value: unknown): value is CharacterListRecord & { id: string } {
   return Boolean(
@@ -590,164 +556,6 @@ export function useDeleteCharacterGalleryImage(characterId: string) {
   });
 }
 
-// ── Personas ──
-
-export function usePersonas(enabled = true) {
-  return useQuery({
-    queryKey: characterKeys.personas,
-    queryFn: () => storageApi.list<unknown>("personas"),
-    enabled,
-    staleTime: 5 * 60_000,
-    refetchOnWindowFocus: false,
-  });
-}
-
-export function usePersonaSummaries(enabled = true) {
-  return useQuery({
-    queryKey: characterKeys.personaSummaries,
-    queryFn: () => storageApi.list<PersonaSummary>("personas", PERSONA_SUMMARY_OPTIONS),
-    enabled,
-    staleTime: 5 * 60_000,
-    refetchOnWindowFocus: false,
-  });
-}
-
-export function usePersona(id: string | null, enabled = true) {
-  return useQuery({
-    queryKey: characterKeys.personaDetail(id ?? ""),
-    queryFn: () => storageApi.get("personas", id!),
-    enabled: enabled && !!id,
-    staleTime: 5 * 60_000,
-    refetchOnWindowFocus: false,
-  });
-}
-
-export function useActivePersona(enabled = true) {
-  return useQuery({
-    queryKey: characterKeys.activePersona,
-    queryFn: async () => {
-      const personas = await storageApi.list<PersonaSummary & { active?: string | boolean }>(
-        "personas",
-        PERSONA_SUMMARY_OPTIONS,
-      );
-      return (
-        personas.find(
-          (persona) =>
-            persona.isActive === true ||
-            persona.isActive === "true" ||
-            persona.active === true ||
-            persona.active === "true",
-        ) ?? null
-      );
-    },
-    enabled,
-    staleTime: 5 * 60_000,
-    refetchOnWindowFocus: false,
-  });
-}
-
-export function useCreatePersona() {
-  const qc = useQueryClient();
-  return useMutation({
-    mutationFn: (data: {
-      name: string;
-      description?: string;
-      comment?: string;
-      personality?: string;
-      scenario?: string;
-      backstory?: string;
-      appearance?: string;
-      nameColor?: string;
-      dialogueColor?: string;
-      boxColor?: string;
-      trackerCardColors?: string;
-      personaStats?: unknown;
-      altDescriptions?: unknown[];
-      tags?: string[];
-      savedStatusOptions?: string[];
-      avatarCrop?: unknown;
-    }) => storageApi.create("personas", data),
-    onSuccess: () => {
-      qc.invalidateQueries({ queryKey: characterKeys.personas });
-      qc.invalidateQueries({ queryKey: characterKeys.personaSummaries });
-    },
-  });
-}
-
-export function useUpdatePersona() {
-  const qc = useQueryClient();
-  return useMutation({
-    mutationFn: ({
-      id,
-      ...data
-    }: {
-      id: string;
-      name?: string;
-      comment?: string;
-      description?: string;
-      personality?: string;
-      scenario?: string;
-      backstory?: string;
-      appearance?: string;
-      nameColor?: string;
-      dialogueColor?: string;
-      boxColor?: string;
-      trackerCardColors?: string;
-      personaStats?: unknown;
-      altDescriptions?: unknown[];
-      tags?: string[];
-      savedStatusOptions?: string[];
-      avatarCrop?: unknown;
-    }) => storageApi.update("personas", id, data),
-    onSuccess: (updatedPersona, variables) => {
-      qc.setQueryData(characterKeys.personaDetail(variables.id), updatedPersona);
-      qc.setQueryData<unknown[] | undefined>(characterKeys.personas, (old) => {
-        if (!Array.isArray(old)) return old;
-        const updatedId = (updatedPersona as { id?: string } | null)?.id ?? variables.id;
-        if (!updatedId) return old;
-
-        return old.map((p) => {
-          const row = p as Record<string, unknown> & { id?: string };
-          if (row?.id !== updatedId) return p;
-          if (!updatedPersona || typeof updatedPersona !== "object") return p;
-          return { ...row, ...(updatedPersona as Record<string, unknown>) };
-        });
-      });
-
-      qc.invalidateQueries({ queryKey: characterKeys.personas });
-      qc.invalidateQueries({ queryKey: characterKeys.personaSummaries });
-      qc.invalidateQueries({ queryKey: characterKeys.personaDetail(variables.id) });
-    },
-  });
-}
-
-export function useDeletePersona() {
-  const qc = useQueryClient();
-  return useMutation({
-    mutationFn: (id: string) => storageApi.delete("personas", id),
-    onSuccess: (_data, id) => {
-      qc.removeQueries({ queryKey: characterKeys.personaDetail(id) });
-      qc.removeQueries({ queryKey: characterKeys.personaSummaryDetail(id) });
-      qc.invalidateQueries({ queryKey: characterKeys.personas });
-      qc.invalidateQueries({ queryKey: characterKeys.personaSummaries });
-    },
-  });
-}
-
-export function useUploadPersonaAvatar() {
-  const qc = useQueryClient();
-  return useMutation({
-    mutationFn: ({ id, avatar, filename }: { id: string; avatar: string; filename?: string }) =>
-      personaApi.uploadAvatar(id, avatar, filename),
-    onSuccess: (_data, variables) => {
-      qc.invalidateQueries({ queryKey: characterKeys.personas });
-      qc.invalidateQueries({ queryKey: characterKeys.personaSummaries });
-      qc.invalidateQueries({ queryKey: characterKeys.personaDetail(variables.id) });
-      qc.invalidateQueries({ queryKey: characterKeys.personaSummaryDetail(variables.id) });
-    },
-  });
-}
-
 // ── Character Groups ──
 
 export function useCharacterGroups() {
@@ -780,15 +588,5 @@ export function useDeleteGroup() {
   return useMutation({
     mutationFn: (id: string) => storageApi.delete("character-groups", id),
     onSuccess: () => qc.invalidateQueries({ queryKey: characterKeys.groups }),
-  });
-}
-
-// ── Persona Groups ──
-
-export function usePersonaGroups(enabled = true) {
-  return useQuery({
-    queryKey: characterKeys.personaGroups,
-    queryFn: () => storageApi.list<unknown>("persona-groups"),
-    enabled,
   });
 }
