+++ src/features/catalog/characters/query-keys.ts
@@ -8,14 +8,7 @@ export const characterKeys = {
   detail: (id: string) => [...characterKeys.all, "detail", id] as const,
   versions: (id: string) => [...characterKeys.detail(id), "versions"] as const,
   gallery: (id: string) => [...characterKeys.all, "gallery", id] as const,
-  personas: ["personas"] as const,
-  personaSummaries: ["personas", "summaries"] as const,
-  personaSummaryDetail: (id: string) => ["personas", "summaries", id] as const,
-  personaDetail: (id: string) => [...characterKeys.personas, "detail", id] as const,
-  activePersona: ["personas", "active"] as const,
   groups: ["character-groups"] as const,
-  personaGroups: ["persona-groups"] as const,
-  personaGroupDetail: (id: string) => ["persona-groups", "detail", id] as const,
 };
 
 export const spriteKeys = {
