+++ src/features/catalog/personas/components/PersonaMakerModal.tsx
@@ -5,7 +5,7 @@
 import { useState, useRef, useCallback } from "react";
 import { Modal } from "../../../../shared/components/ui/Modal";
 import { useConnections } from "../../connections/index";
-import { useCreatePersona } from "../../characters/index";
+import { useCreatePersona } from "../hooks/use-personas";
 import { useUIStore } from "../../../../shared/stores/ui.store";
 import { Sparkles, Loader2, Wand2, CheckCircle, AlertCircle, ChevronDown, User, Save } from "lucide-react";
 import { ProfessorMariWorkingWindow } from "../../../../shared/components/ui/ProfessorMariWorkingWindow";
