+++ src/features/catalog/personas/index.ts
@@ -1,0 +1,2 @@
+export * from "./query-keys";
+export * from "./hooks/use-personas";
