+++ src/features/catalog/personas/query-keys.ts
@@ -1,4 +1,9 @@
 export const personaKeys = {
   list: ["personas"] as const,
+  summaries: ["personas", "summaries"] as const,
+  summaryDetail: (id: string) => ["personas", "summaries", id] as const,
+  detail: (id: string) => [...personaKeys.list, "detail", id] as const,
+  active: ["personas", "active"] as const,
   groups: ["persona-groups"] as const,
+  groupDetail: (id: string) => ["persona-groups", "detail", id] as const,
 };
